% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spiderman.R
\name{spiderman}
\alias{spiderman}
\title{Spider-Man}
\usage{
spiderman(people_saved, saved_from, day)
}
\arguments{
\item{people_saved}{total number of people Spider-man saved}

\item{saved_from}{what did Spider-man save the people from?}

\item{day}{Was this just a regular old Tuesday?}
}
\description{
Spider-Man
}
\examples{
spiderman(9, "Venom", "a regular old Tuesday")


}
